+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.0
+### Added
+* `SliverStickyHeaderState` class. This state has `scrollPercentage` and `isPinned` properties and it's passed to the `SliverStickyHeaderBuilder` when it changes.
+
+### Changed
+* The second parameter for `SliverStickyHeaderWidgetBuilder` takes now a `SliverStickyHeaderState` instead of a double.
+
 ## 0.2.0
 ### Removed
 * `sliverStickyHeaderScrollNotifier` argument in `SliverStickyHeader` constructor.